Refactoring SDK reflaction utils into core.sdk
