+++ src/command_modules/azure-cli-monitor/azure/cli/command_modules/monitor/validators.py
@@ -7,25 +7,6 @@ from azure.cli.core.commands.arm import is_valid_resource_id, resource_id, parse
 from azure.cli.core._util import CLIError
 
 
-def get_complex_argument_processor(expanded_arguments, assigned_arg, model_type):
-    """
-    Return a validator which will aggregate multiple arguments to one complex argument.
-    """
-    def _expansion_validator_impl(namespace):
-        """
-        The validator create a argument of a given type from a specific set of arguments from CLI
-        command.
-        :param namespace: The argparse namespace represents the CLI arguments.
-        :return: The argument of specific type.
-        """
-        ns = vars(namespace)
-        kwargs = dict((k, ns[k]) for k in ns if k in set(expanded_arguments))
-
-        setattr(namespace, assigned_arg, model_type(**kwargs))
-
-    return _expansion_validator_impl
-
-
 # pylint: disable=line-too-long
 def validate_diagnostic_settings(namespace):
     from azure.cli.core.commands.client_factory import get_subscription_id
